backend/requirements.txt
# Core API framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# OpenAI client for vLLM API
openai==1.54.0
httpx==0.27.2

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Image processing
Pillow==10.2.0
pdf2image==1.17.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Logging and utilities
python-json-logger==2.0.7

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0
